fix(food): split PRD-133 log-inference under the 200-line file cap
CI lint flagged eslint(max-lines) at 204 lines on `log-inference.ts`
after the cost-cap + default-sink additions. Split by responsibility:

- `log-inference-types.ts` (81L) — shared input/opts/deps types.
- `log-inference-sink.ts` (41L) — the env-driven default sink
  `logFoodInference` (POST to `food.ai.logInference`).
- `log-inference.ts` (187L) — the `callClaudeWithLogging` wrapper +
  type re-exports so the documented entrypoint
  `@pops/app-food/src/ai/log-inference` keeps the same public API.

No behaviour change; same 17 tests pass.
